fix(amenable_verus): compare via a widening cast, not a narrowing one
converted == value as u8 (i32 as u8, flagged by clippy::cast_sign_loss)
became converted as i32 == value -- u8 as i32 is a widening
unsigned-to-strictly-wider-signed cast, which can never lose sign
information (there is none to lose) and which clippy never flags.

First attempt used i32::from(converted) (a lossless From conversion,
avoiding `as` entirely) -- reverted after real cargo verus rejected it:
`core::convert::num::impl&%87::from is not supported`, since vstd has no
spec for that specific From impl, unlike its existing TryFrom<i32> for
u8 coverage this file's own header comment already relies on. The `as`
cast form is a primitive Verus understands directly, not a trait method
needing an explicit vstd spec, and real verus confirms it: 491 verified,
0 errors.

No #[allow] anywhere -- a real fix existed once the cast direction was
reconsidered, not just the first (broken) alternative tried.

+++ crates/amenable_verus/src/rust_std/try_from_int_error_carrier.rs
@@ -36,7 +36,7 @@ pub fn verify_try_from_int_error_occurs_exactly_when_out_of_range(value: i32) ->
         result,
 {
     match <u8 as std::convert::TryFrom<i32>>::try_from(value) {
-        Ok(converted) => (0 <= value && value <= u8::MAX as i32) && converted == value as u8,
+        Ok(converted) => (0 <= value && value <= u8::MAX as i32) && converted as i32 == value,
         Err(_) => value < 0 || value > u8::MAX as i32,
     }
 }
